better makefile
