+++ firmware/test/testcommon.h
@@ -18,11 +18,13 @@ uint32_t rand32(void);
 
 extern __xdata USB_HUB_PORT UsbDev;
 
-void TestSetup();
-uint8_t * GetNextChunk();
-uint8_t * GetNextChonk();
+void TestSetup(void);
+uint8_t * GetNextChunk(void);
+uint8_t * GetNextChonk(void);
 
-void halt();
+void halt(void);
+
+extern __code uint32_t bitMasks32[];
 
 /* Debugging enabled -- verify assertions at run time. */
 void __assert(const char *expression, const char *functionname, const char *filename, unsigned int linenumber);
