+++ firmware/test/testendtoend.c
@@ -37,79 +37,6 @@
 __xdata USB_HUB_PORT UsbDev;
 
 
-
-bool TestExtractValue(uint8_t reportSize, uint16_t startBit){
-
-    static __xdata HID_SEG testSeg;
-    
-    testSeg.startBit = startBit;
-    testSeg.reportSize = reportSize;
-
-    //DumpHID(pInterface);
-
-    static __xdata uint32_t ogtestdata;
-    static __xdata uint32_t testdata;
-    static __xdata uint32_t afterdata;
-
-    //start with all zeros
-    testdata = 0;
-    afterdata = SegExtractValue(&testSeg, (__xdata uint8_t *)(&testdata));
-    assert(testdata == afterdata);
-
-    //now all 1s
-    ogtestdata = (0xFFFFFFFF & bitMasks32[reportSize]);
-    testdata = ogtestdata << startBit;
-    afterdata = SegExtractValue(&testSeg, (__xdata uint8_t *)(&testdata));
-    assert(ogtestdata == afterdata);
-
-    //now a few random values
-    for (uint8_t c = 0; c < 20; c++) {
-        ogtestdata = rand32() & bitMasks32[reportSize];
-        testdata = ogtestdata << startBit;
-        afterdata = SegExtractValue(&testSeg, (__xdata uint8_t *)(&testdata));
-        assert(ogtestdata == afterdata);
-    }
-
-    // if we're testing a single bit, set JUST that bit and make sure none of the other bits return 1
-    if (reportSize == 1) {
-        testdata = (uint32_t)0x01 << (uint32_t)startBit;
-
-        for (uint8_t d = 0; d < 32; d++){
-            testSeg.startBit = d;
-
-            // the bit we're looking for should be 1
-            if (d == startBit) {
-
-                assert (SegExtractValue(&testSeg, (__xdata uint8_t *)(&testdata)) == 0x01);
-            }
-
-            // the bits we're not looking for should be zero
-            else {
-                assert (SegExtractValue(&testSeg, (__xdata uint8_t *)(&testdata)) != 0x01);
-            }
-        }
-    }
-
-    return 1;
-}
-
-void TestSegExtracts(){
-    for (int size=1; size <= 32; size++) {
-        for (int sb=0; sb <= 32 - size; sb++){
-            //printf ("%d-%d ", size, sb);
-            assert(TestExtractValue(size, sb));
-        }
-    }
-
-    printf("HID Segment Offset/Size Test Passed\n");
-}
-
-
-
-
-
-
-
 void main()
 {
     // timer0 setup
